backend/app/utils/remote_db.py
"""Utilities for querying remote databases: list tables, list columns, fetch sample data, compute structure hash."""

from __future__ import annotations
import hashlib
import json
from typing import List, Tuple, Optional, Any, Dict

from app.utils.db_connector import test_connection


def _get_mysql_conn(host, port, user, password, database, charset="utf8", timeout=10):
    import pymysql
    return pymysql.connect(
        host=host, port=port, user=user, password=password,
        database=database or "", charset=charset or "utf8",
        connect_timeout=timeout,
    )


def _get_pg_conn(host, port, user, password, database, timeout=10):
    import psycopg2
    dsn = f"host={host} port={port} user={user} password={password} connect_timeout={timeout}"
    if database:
        dsn += f" dbname={database}"
    return psycopg2.connect(dsn)


def _get_sqlserver_conn(host, port, user, password, database, timeout=10):
    import pyodbc
    conn_str = (
        f"DRIVER={{ODBC Driver 17 for SQL Server}};"
        f"SERVER={host},{port};"
        f"UID={user};PWD={password};"
        f"CONNECTION TIMEOUT={timeout};"
    )
    if database:
        conn_str += f"DATABASE={database};"
    return pyodbc.connect(conn_str, timeout=timeout)


def _connect(db_type, host, port, user, password, database=None, schema=None, charset="utf8", timeout=10):
    if db_type == "mysql":
        return _get_mysql_conn(host, port, user, password, database, charset, timeout)
    elif db_type == "postgresql":
        return _get_pg_conn(host, port, user, password, database, timeout)
    elif db_type == "sqlserver":
        return _get_sqlserver_conn(host, port, user, password, database, timeout)
    else:
        raise ValueError(f"不支持的数据库类型: {db_type}")


def list_tables(db_type: str, host: str, port: int, user: str, password: str,
                database: str | None = None, schema: str | None = None,
                charset: str = "utf8", timeout: int = 10) -> List[Dict]:
    """Return list of {table_name, table_comment}."""
    conn = _connect(db_type, host, port, user, password, database, schema, charset, timeout)
    try:
        cur = conn.cursor()
        if db_type == "mysql":
            sql = "SELECT TABLE_NAME, TABLE_COMMENT FROM information_schema.TABLES WHERE TABLE_SCHEMA = %s AND TABLE_TYPE = 'BASE TABLE' ORDER BY TABLE_NAME"
            cur.execute(sql, (database or "",))
        elif db_type == "postgresql":
            sch = schema or "public"
            sql = """
                SELECT c.relname AS table_name,
                       obj_description(c.oid) AS table_comment
                FROM pg_class c
                JOIN pg_namespace n ON n.oid = c.relnamespace
                WHERE n.nspname = %s AND c.relkind = 'r'
                ORDER BY c.relname
            """
            cur.execute(sql, (sch,))
        elif db_type == "sqlserver":
            sch = schema or "dbo"
            sql = """
                SELECT t.name AS table_name,
                       ep.value AS table_comment
                FROM sys.tables t
                LEFT JOIN sys.extended_properties ep
                    ON ep.major_id = t.object_id AND ep.minor_id = 0 AND ep.name = 'MS_Description'
                LEFT JOIN sys.schemas s ON s.schema_id = t.schema_id
                WHERE s.name = ?
                ORDER BY t.name
            """
            cur.execute(sql, (sch,))

        rows = cur.fetchall()
        result = []
        for r in rows:
            result.append({
                "table_name": r[0],
                "table_comment": r[1] if len(r) > 1 and r[1] else None,
            })
        return result
    finally:
        conn.close()


def list_columns(db_type: str, host: str, port: int, user: str, password: str,
                 table_name: str, database: str | None = None, schema: str | None = None,
                 charset: str = "utf8", timeout: int = 10) -> List[Dict]:
    """Return list of column dicts with keys: field_name, db_data_type, is_nullable, column_default, is_primary_key, ordinal_position."""
    conn = _connect(db_type, host, port, user, password, database, schema, charset, timeout)
    try:
        cur = conn.cursor()
        if db_type == "mysql":
            sql = """
                SELECT c.COLUMN_NAME, c.COLUMN_TYPE, c.IS_NULLABLE, c.COLUMN_DEFAULT, c.ORDINAL_POSITION,
                       CASE WHEN kcu.COLUMN_NAME IS NOT NULL THEN 1 ELSE 0 END AS is_pk
                FROM information_schema.COLUMNS c
                LEFT JOIN information_schema.KEY_COLUMN_USAGE kcu
                    ON kcu.TABLE_SCHEMA = c.TABLE_SCHEMA
                    AND kcu.TABLE_NAME = c.TABLE_NAME
                    AND kcu.COLUMN_NAME = c.COLUMN_NAME
                    AND kcu.CONSTRAINT_NAME = 'PRIMARY'
                WHERE c.TABLE_SCHEMA = %s AND c.TABLE_NAME = %s
                ORDER BY c.ORDINAL_POSITION
            """
            cur.execute(sql, (database or "", table_name))
        elif db_type == "postgresql":
            sch = schema or "public"
            sql = """
                SELECT c.column_name, c.data_type, c.is_nullable, c.column_default, c.ordinal_position,
                       CASE WHEN tc.constraint_type = 'PRIMARY KEY' THEN 1 ELSE 0 END AS is_pk
                FROM information_schema.columns c
                LEFT JOIN information_schema.key_column_usage kcu
                    ON kcu.table_schema = c.table_schema AND kcu.table_name = c.table_name AND kcu.column_name = c.column_name
                LEFT JOIN information_schema.table_constraints tc
                    ON tc.constraint_name = kcu.constraint_name AND tc.table_schema = kcu.table_schema AND tc.constraint_type = 'PRIMARY KEY'
                WHERE c.table_schema = %s AND c.table_name = %s
                ORDER BY c.ordinal_position
            """
            cur.execute(sql, (sch, table_name))
        elif db_type == "sqlserver":
            sch = schema or "dbo"
            sql = """
                SELECT c.name AS column_name,
                       TYPE_NAME(c.user_type_id) + CASE WHEN c.max_length > 0 AND TYPE_NAME(c.user_type_id) IN ('varchar','nvarchar','char','nchar') THEN '(' + CAST(c.max_length AS VARCHAR) + ')' ELSE '' END AS data_type,
                       c.is_nullable,
                       dc.definition AS column_default,
                       c.column_id AS ordinal_position,
                       CASE WHEN ic.column_id IS NOT NULL THEN 1 ELSE 0 END AS is_pk
                FROM sys.columns c
                JOIN sys.tables t ON t.object_id = c.object_id
                JOIN sys.schemas s ON s.schema_id = t.schema_id
                LEFT JOIN sys.default_constraints dc ON dc.object_id = c.default_object_id
                LEFT JOIN sys.indexes i ON i.object_id = t.object_id AND i.is_primary_key = 1
                LEFT JOIN sys.index_columns ic ON ic.object_id = i.object_id AND ic.index_id = i.index_id AND ic.column_id = c.column_id
                WHERE s.name = ? AND t.name = ?
                ORDER BY c.column_id
            """
            cur.execute(sql, (sch, table_name))

        rows = cur.fetchall()
        result = []
        for r in rows:
            result.append({
                "field_name": r[0],
                "db_data_type": str(r[1]),
                "is_nullable": r[2] in (True, 'YES', 1),
                "column_default": str(r[3]) if r[3] is not None else None,
                "ordinal_position": r[4],
                "is_primary_key": bool(r[5]),
            })
        return result
    finally:
        conn.close()


def fetch_sample_data(db_type: str, host: str, port: int, user: str, password: str,
                      table_name: str, database: str | None = None, schema: str | None = None,
                      charset: str = "utf8", timeout: int = 10,
                      limit: int = 5) -> Tuple[List[str], List[List[Optional[str]]]]:
    """Return (columns, rows) where each cell is a string."""
    conn = _connect(db_type, host, port, user, password, database, schema, charset, timeout)
    try:
        cur = conn.cursor()
        qualified = table_name
        if db_type == "postgresql":
            sch = schema or "public"
            qualified = f'"{sch}"."{table_name}"'
        elif db_type == "sqlserver":
            sch = schema or "dbo"
            qualified = f"[{sch}].[{table_name}]"

        if db_type == "sqlserver":
            cur.execute(f"SELECT TOP {limit} * FROM {qualified}")
        else:
            cur.execute(f"SELECT * FROM {qualified} LIMIT {limit}")

        columns = [desc[0] for desc in cur.description]
        raw_rows = cur.fetchall()
        rows = []
        for raw in raw_rows:
            rows.append([str(v) if v is not None else None for v in raw])
        return columns, rows
    finally:
        conn.close()


def compute_structure_hash(columns: List[Dict]) -> str:
    """Compute a deterministic hash from column list for structure change detection."""
    normalized = []
    for c in columns:
        normalized.append({
            "field_name": c["field_name"],
            "db_data_type": c["db_data_type"],
            "is_primary_key": c.get("is_primary_key", False),
            "ordinal_position": c.get("ordinal_position", 0),
        })
    normalized.sort(key=lambda x: x["ordinal_position"])
    payload = json.dumps(normalized, sort_keys=True, ensure_ascii=False)
    return hashlib.sha256(payload.encode("utf-8")).hexdigest()[:32]
