backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pymysql==1.1.1
psycopg2-binary==2.9.9
pyodbc==5.1.0
openpyxl==3.1.5
